build QuizProvider.js to provide context to all app
- move state from App component to QuizProvider
- build custom hook inside the custome provider component
- suround the app with the custom context provider
